tarefas-projeto/backed/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
PyMySQL==1.0.3
python-dotenv==1.0.0